+++ backend/src/main/java/ch/rfobaden/incidentmanager/backend/services/ReportService.java
@@ -34,15 +34,14 @@ public class ReportService extends ModelRepositoryService<Report, ReportPath, Re
 
     @Override
     public Optional<Report> update(ReportPath path, Report report) {
-        Optional<Report> oldReport = repository.findById(report.getId());
+        Optional<Report> oldReport = find(path, report.getId());
         if (oldReport.isEmpty()) {
             return Optional.empty();
         }
-        Long oldReportId = oldReport.get().getAssigneeId();
         Optional<Report> savedReport = super.update(path, report);
         savedReport.ifPresent(rep -> {
             if (rep.getAssigneeId() != null
-                && !rep.getAssigneeId().equals(oldReportId)) {
+                && !rep.getAssigneeId().equals(oldReport.get().getAssigneeId())) {
                 sendAssignmentEmail(rep);
             }
         });