+++ frontend/components/Ui/Container/UiContainer.tsx
@@ -1,6 +1,5 @@
 import { ReactNode } from 'react'
 import styled, { css } from 'styled-components'
-import StyleHelper from '@/utils/helpers/StyleHelper'
 import { Themed } from '@/theme'
 
 interface Props {
@@ -8,25 +7,32 @@ interface Props {
   children?: ReactNode
 }
 
-const UiContainer = styled(StyleHelper.tag<Props>('div'))`
+const UiContainer = styled.div<Props>`
+  ${({ isFluid }) => !isFluid ? fluidCss : sharedCss}
+`
+
+const sharedCss = css`
   width: 100%;
   margin-inline: auto;
-  ${({ isFluid }) => !isFluid && css`
-    padding: 0 0.8rem;
-    ${Themed.media.sm.min} {
-      max-width: ${({ theme }) => theme.breakpoints.sm.min}px;
-    }
-    ${Themed.media.md.min} {
-      max-width: ${({ theme }) => theme.breakpoints.md.min}px;
-    }
-    ${Themed.media.lg.min} {
-      padding: 0 2rem;
-      max-width: ${({ theme }) => theme.breakpoints.lg.min}px;
-    }
-    ${Themed.media.xl.min} {
-      padding: 0 4rem;
-      max-width: ${({ theme }) => theme.breakpoints.xl.min}px;
-    }
-  `}
 `
-export default UiContainer
+
+const fluidCss = css`
+  ${sharedCss};
+  
+  padding-left: 0.8rem;
+  padding-right: 0.8rem;
+  
+  ${Themed.media.sm.min} {
+    max-width: ${({ theme }) => theme.breakpoints.sm.min}px;
+  }
+  ${Themed.media.md.min} {
+    max-width: 100%;
+    
+    padding-right: 4rem;
+    padding-left: 4rem;
+  }
+`
+
+export default Object.assign(UiContainer, {
+  fluidCss,
+})