+++ frontend/components/Ui/DropDown/Trigger/UiDropDownTrigger.tsx
@@ -1,0 +1,14 @@
+import React, { ReactNode } from 'react'
+
+interface Props {
+  children: ReactNode
+}
+
+const UiDropDownTrigger: React.VFC<Props> = ({ children }) => {
+  return (
+    <React.Fragment>
+      {children}
+    </React.Fragment>
+  )
+}
+export default UiDropDownTrigger