+++ frontend/components/Ui/List/UiList.tsx
@@ -14,8 +14,10 @@ const UiList: React.VFC<Props> = ({ children }) => {
 }
 export default UiList
 
-const Box = styled.div`
+const Box = styled.ul`
+  width: 100%;
+  height: 100%;
   display: flex;
   flex-direction: column;
-  gap: 1rem 0;
+  row-gap: 0.4rem;
 `